make functional latency compensated apriltag positioning

+++ src/main/include/Constants.h
@@ -8,7 +8,7 @@ namespace MainConst {
 }
 
 namespace DriverControllerConstants {
-    inline constexpr float dB = 0.1;
+    inline constexpr double dB = 0.1;
 }
 
 // constants for the swerve
